Update master README from dev branch
Given that the 0.9.5 and 0.9.6 versions of NodeMCU are deprecated, it
is quite important that the updated README appears in the master branch.
That's what people will see when they visit the repository page on github.

This commit just squashes all recent dev-branch changes to README.md
to get it updated ASAP in advance of pulling the entire dev branch
into master.

Signed-off-by: Nick Andrew <nick@nick-andrew.net>