Addition of new task IF with GPIO rework to supprt this
The following files have been updated to change the Flash version of the pin map
into a 32 packed structure. This simplifies the map declaration, but this is
copied to the existing the RAM vector format to minimise coding impact on this
patch iteration. We might swith to the structure version in RM in future.

-  `app/platform/pin_map.h`
-  `app/platform/pin_map.c`

The platform interface has been updated:

-   To move ISR code out of Flash and into RAM to comply with SDK guidelines and
    for optimal execution speed. (Note that you need to disable inlining
    optimization for any such routines.)

-   The ISR bit-test loop has been peephole optimised to shorten the maximum ISR
    time.

-   To split our Lua land from the Hardware abstraction layer. The hardware
    knows nothing about Lua or the Lua runtime environent, so the platform
    modules now make no reference to Lua. A new vector, pin_trigger, is used to
    flag if a callback has been attached to a GPIO pin and in which case the ISR
    posts a task to dispatch this. The registering of any Lua callbacks, and
    their unregistering is carried out within the Lua libraries.

-   As a result of this platform_gpio_mode() no longer unreferences any lua
    callback, so this must be done in the calling layer.

The Lua Library module gpio.c has also been updated to use the new task and
updated platform laver.  Since I've been working these routines, I have also
tidied up their implementation to use the standard Lua approach to argument
manipulation, and this has slightly reduced the code size as well as making
the implementaiton a lot more readable.

This PR also includes the previously discussed task interface and changes to
the UART driver to support this.

I have also added a draft Extension Developer FAQ which decribes how the SDK
implements ISRs and tasks.