fix unexpected low level after writing
first commit to fix issue #1130

If power = 0, after writing the bus is pulled down, and the slave will take it for a bit. One may have problem like the onewire device returning strange response, (for me, ds18b20 sometimes give temperature 4095.8125 degc). This commit will fix the unexpected situation. 

The power argument is not removed. We believe some use this module without external pull up as the documentation specified. The direct push-pull should be able to drive the slave. In that case, we may not switch the pin to input mode for the push arm supplying the power.