Major cleanup - c_whatever is finally history.
This was a royal pain to clean up, but I believe it's now all done.
If Espressif had deigned to ship the corresponding libc headers this
wouldn't have been needed in the first place, but oh well.

Essentially, the libc.a provided by the SDK uses non-standard memory
allocation, so anything that would use the regular malloc()/free()
ends up looking for things that do not exist. Why they didn't provide
wrappers for the whole lot is beyond me. Even worse is the libc.a they
ship includes functions which rely on the missing alloc routines!

Anyway, app/libc has now been cut down to the bare minimum overrides to
shadow the corresponding functions in the SDK's libc. The old c_xyz.h
headerfiles have been nuked in favour of the standard <xyz.h> headers,
with a few exceptions over in sdk-overrides. Again, shipping a libc.a
without headers is a terrible thing to do. We're still living on a
prayer that libc was configured the same was as a default-configured
xtensa gcc toolchain assumes it is. That part I cannot do anything about,
unfortunately, but it's no worse than it has been before.

The linker ordering in app/Makefile has been updated to ensure we
link in our overrides first, without needing to rely on the crutch
of prefixing everything with c_ (yuck!).

From here on out it's now possible (and needed) to include the standard
header files, and use the typical malloc()/calloc()/realloc()/free(),
the strwhatever()s and memwhatever()s. These end up, through macro
and linker magic, mapped to the appropriate SDK or ROM functions.

I'm still impressed the NodeMCU founders managed to get things going,
considering the state of the SDK...