Issue restart rather than second lua_main() call on LFS reload.
By the time we get informed that there's an LFS reload we've already
opened all the user libraries as well, so it's not safe to simply redo
that with a new lua_State. Currently that results in a hang, and a
manual reset is required to progress. Doing the automatic restart is
the correct approach here, even though this adds another reboot to the
LFS reload process.

Also fixed up a couple of minor issues when embedded-LFS is used.