Fix failure to close listening socket (#3679)
The IDF LWIP's `netconn_close()` no longer returns valid error codes as the #ifdef'd code path the IDF uses does not end up actually doing much, which unfortunately includes actually setting the error code. As such, checking the return code is actively harmful to proper operation for us. The fix here is simply to carry on with the `netconn_delete()`. This _should_ be safe even if `netconn_close()` actually errored.